thermal: cpu_cooling: Remove pointless dependency on CONFIG_OF
The option CONFIG_CPU_THERMAL depends on CONFIG_OF in the Kconfig.

It it pointless to check if CONFIG_OF is set in the header file as
this is always true if CONFIG_CPU_THERMAL is true. Remove it.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Acked-by: Viresh Kumar <viresh.kumar@linaro.org>
Reviewed-by: Amit Kucheria <amit.kucheria@linaro.org>
Link: https://lore.kernel.org/r/20191030091038.678-1-daniel.lezcano@linaro.org
Signed-off-by: Divyanshu-Modi <divyan.m05@gmail.com>
(cherry picked from commit 3e6a8adce60e199b5f7b207e14f9b9828fb42cbb)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>