crypto: x86/chacha - depend on generic chacha library instead of crypto driver
commit 28e8d89b1ce8d2e7badfb5f69971dd635acb8863 upstream.

In preparation of extending the x86 ChaCha driver to also expose the ChaCha
library interface, drop the dependency on the chacha_generic crypto driver
as a non-SIMD fallback, and depend on the generic ChaCha library directly.
This way, we only pull in the code we actually need, without registering
a set of ChaCha skciphers that we will never use.

Since turning the FPU on and off is cheap these days, simplify the SIMD
routine by dropping the per-page yield, which makes for a cleaner switch
to the library API as well. This also allows use to invoke the skcipher
walk routines in non-atomic mode.

Signed-off-by: Ard Biesheuvel <ardb@kernel.org>
Change-Id: I247374d6ba0ca89eef3859b1a5afaf42b13f61e0
Signed-off-by: Herbert Xu <herbert@gondor.apana.org.au>
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
(cherry picked from commit e26ad65e01d68816d6d75a9f95ec0957f763ac6c)
Signed-off-by: TogoFire <togofire@mailfence.com>