kernel/cpu: Silence abundance of logspam
We don't really need to know if the CPU is getting disabled or enabled on a production device.

[Kanishk]: This kernel will explicitly, and always be used in Production builds. It is not needed to keep the logging enabled
each time a CPU core is taken offline and/or sent to deep sleep or woken up. It causes more log processes
to trigger, leading to increased power consumption as well as CPU overhead.

Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
Change-Id: Ide77c2a638bbc74a052530bb49e5ef2d708c59f0
Signed-off-by: Kanishk <kanishkthederp@gmail.com>
Signed-off-by: TogoFire <togofire@mailfence.com>
Signed-off-by: maxsteeel <109047395+maxsteeel@users.noreply.github.com>