thaw_fingerprintd: Use strncmp() instead of memcmp()
memcmp() requires specific filename, and the fingerprintd hal can be upgraded to a newer version which will require us to modify the code again, hence let's switch to strncmp().

This makes the code more efficient in a way as we're using sizeof() which is determined at compile-time, so the compiler will calculate the size of the string and make it a constant.

Suggested-by: Justin Crawford <Justin@stacksmash.net>
Signed-off-by: Justin Crawford <Justin@stacksmash.net>
Signed-off-by: Cyber Knight <cyberknight755@gmail.com>
(cherry picked from commit 050fe5f4bc7863c361d286fc33da820741c0db75)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>