arm64: dts: Don't blank screen as part of LP1 cmd
We send the LP1 command as part of entering Always-On Display mode, but
we don't want that command to blank the screen. After all, the point of
AoD mode is to keep the screen on to display the time, notifications,
etc.

Signed-off-by: chandu078 <chandudyavanapelli03@gmail.com>
Signed-off-by: Pranav Vashi <neobuddy89@gmail.com>
(cherry picked from commit d4e4d924a73035a6461e0f7a92a3c0b140a4d26b)
Signed-off-by: Madara273 <ravenhoxs@gmail.com>